default name is better, kept the previsou one for backward compatibility
